pyproject.toml
[project]
name = "backend-assignement"
version = "0.1.0"
description = ""
authors = [
    {name = "Affaan Jaweed",email = "affaan190@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "bcrypt (==4.0.1)",
    "dotenv (>=0.9.9,<0.10.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
